+++ Tests/YAJLParserTest.m
@@ -7,18 +7,14 @@
 //
 
 
-@interface YAJLParserTest : GHTestCase {}
+@interface YAJLParserTest : YAJLTestCase {}
 @end
 
 @implementation YAJLParserTest
 
 - (void)test {	
-	
-	NSString *examplePath = [[NSBundle mainBundle] pathForResource:@"example" ofType:@"json"];
-	NSData *testData = [[NSData dataWithContentsOfFile:examplePath options:NSUncachedRead error:nil] retain];
-	
-	YAJLParser *parser = [[YAJLParser alloc] initWithData:testData parserOptions:0];
-	[parser parse];
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:0];
+	[parser parse:[self loadData:@"example"]];
 	
 	NSError *error = [parser parserError];
 	if (error) {
@@ -27,17 +23,12 @@
 	}
 	
 	[parser release];
-	[testData release];
 }
 
 - (void)testError {
-	
-	NSString *path = [[NSBundle mainBundle] pathForResource:@"error" ofType:@"json"];
-	NSData *data = [[NSData dataWithContentsOfFile:path options:NSUncachedRead error:nil] retain];
-	
-	YAJLParser *parser = [[YAJLParser alloc] initWithData:data parserOptions:0];
-	BOOL ok = [parser parse];
-	GHAssertFalse(ok, @"Should not have OK status");
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:0];
+	YAJLParserStatus status = [parser parse:[self loadData:@"error"]];
+	GHAssertTrue(status == YAJLParserStatusError, @"Should have error status");
 	
 	NSError *error = [parser parserError];
 	if (error) {
@@ -47,16 +38,12 @@
 	}
 	
 	[parser release];
-	[data release];
 }
 
 - (void)testComments {
-	
-	NSString *path = [[NSBundle mainBundle] pathForResource:@"comments" ofType:@"json"];
-	NSData *data = [[NSData dataWithContentsOfFile:path options:NSUncachedRead error:nil] retain];
-	
-	YAJLParser *parser = [[YAJLParser alloc] initWithData:data parserOptions:YAJLParserOptionsAllowComments];
-	[parser parse];
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:YAJLParserOptionsAllowComments];
+	YAJLParserStatus status = [parser parse:[self loadData:@"comments"]];
+	GHAssertTrue(status == YAJLParserStatusOK, @"Should have error status");
 	
 	NSError *error = [parser parserError];
 	if (error) {
@@ -65,16 +52,12 @@
 	}
 	
 	[parser release];
-	[data release];
 }
 
 - (void)testFailOnComments {
-	
-	NSString *path = [[NSBundle mainBundle] pathForResource:@"comments" ofType:@"json"];
-	NSData *data = [[NSData dataWithContentsOfFile:path options:NSUncachedRead error:nil] retain];
-	
-	YAJLParser *parser = [[YAJLParser alloc] initWithData:data parserOptions:0];
-	[parser parse];
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:0];
+	YAJLParserStatus status = [parser parse:[self loadData:@"comments"]];
+	GHAssertTrue(status == YAJLParserStatusError, @"Should have error status");
 	
 	NSError *error = [parser parserError];
 	if (error) {
@@ -84,7 +67,43 @@
 	}
 
 	[parser release];
-	[data release];
+}
+
+- (void)testOverflow {
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:0];
+	YAJLParserStatus status = [parser parse:[self loadData:@"overflow"]];
+	GHAssertTrue(status == YAJLParserStatusError, @"Should have error status");
+	
+	NSError *error = [parser parserError];
+	if (error) {
+		GHTestLog(@"Parse error:\n%@", error);		
+	} else {
+		GHFail(@"Should have error");
+	}
+	
+	[parser release];
+}
+
+- (void)testStreaming {
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:0];
+	YAJLParserStatus status;
+	status = [parser parse:[self loadData:@"stream1"]];
+	GHAssertTrue(status == YAJLParserStatusInsufficientData, @"Should have insufficient data");
+
+	status = [parser parse:[self loadData:@"stream2"]];
+	GHAssertTrue(status == YAJLParserStatusOK, @"Should have finished");
+	[parser release];
+}
+
+// TODO(gabe): Should error if you try to re-use
+- (void)_testParseMoreThanOnce {
+	YAJLParser *parser = [[YAJLParser alloc] initWithParserOptions:0];
+	parser.delegate = self;
+	YAJLParserStatus status;
+	status = [parser parse:[self loadData:@"example"]];
+	GHAssertTrue(status == YAJLParserStatusOK, @"Should be OK");
+	status = [parser parse:[self loadData:@"example"]];
+	GHAssertTrue(status == YAJLParserStatusError, @"Should error");
 }
 
 @end